+++ terraform/lambda.tf
@@ -1,5 +1,10 @@
 module "lambda" {
   source = "./modules/lambda"
 
-  function_name = "hello-world"
+  function_name        = "helloworld"
+  function_description = "This is a test function for terraform config"
+
+function_tags = {
+  testing = "tested"
+}
 }
