+++ terraform/modules/lambda/iam.tf
@@ -10,33 +10,37 @@ data "aws_iam_policy_document" "assume_lambda_role" {
   }
 }
 
+locals {
+  assume_role_name = "AssumeLambdaRole_${var.function_name}"
+}
+
 resource "aws_iam_role" "lambda" {
-  name = "AssumeLambdaRole"
-  description = "Role for lambda to assume"
+  name               = local.assume_role_name
+  description        = "Role for lambda to assume"
   assume_role_policy = data.aws_iam_policy_document.assume_lambda_role.json
 }
 
 data "aws_iam_policy_document" "allow_lambda_logging" {
-    statement {
-      effect = "Allow"
-      actions = [ 
-        "logs:CreateLogStream",
-        "logs:PutLogEvents",
-      ]
+  statement {
+    effect = "Allow"
+    actions = [
+      "logs:CreateLogStream",
+      "logs:PutLogEvents",
+    ]
 
-      resources = [
-        "arn:aws:logs:*:*:*",
-      ]
-    }
+    resources = [
+      "arn:aws:logs:*:*:*",
+    ]
+  }
 }
 
 resource "aws_iam_policy" "function_logging_policy" {
-  name = "AllowLambdaLoggingPolicy"
+  name        = "AllowLambdaLoggingPolicy_${var.function_name}"
   description = "Policy for lambda cloudwatch logging"
-  policy = data.aws_iam_policy_document.allow_lambda_logging.json
+  policy      = data.aws_iam_policy_document.allow_lambda_logging.json
 }
 
 resource "aws_iam_role_policy_attachment" "lambda_logging_policy_attachment" {
-  role = aws_iam_role.lambda.id
+  role       = aws_iam_role.lambda.id
   policy_arn = aws_iam_policy.function_logging_policy.arn
 }
