+++ terraform/modules/lambda/locals.tf
@@ -1,8 +1,8 @@
 locals {
-  src_path     = "${path.module}/${var.function_name}"
+  src_path = "github.com/spydersweb/aws/lambda/${var.function_name}"
 
-  binary_name  = var.function_name
-  binary_path  = "${path.module}/tf_generated/${local.binary_name}"
+  binary_name  = "bootstrap" #var.function_name
+  binary_path  = "${path.module}/tf_generated"
   archive_path = "${path.module}/tf_generated/${var.function_name}.zip"
 }
